+++ examples/trivial.rs
@@ -1,13 +1,14 @@
 #[dbstruct::dbstruct(db=sled)]
 pub struct Test {
+    #[dbstruct(Default = "Some(42)")]
     the_awnser: Option<u8>,
     the_question: Option<String>,
 }
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let db = Test::open_path("trivial_example")?;
-    db.the_awnser().set(&42)?;
-    db.the_question().set("The Ultimate Question")?;
+    db.the_awnser().set(&Some(42))?;
+    db.the_question().set(Some("The Ultimate Question"))?;
 
     // dropping the db here simulates the program
     // stopping and restarting